Refactored CNN
