+++ .gitignore
@@ -102,3 +102,8 @@ venv.bak/
 
 # mypy
 .mypy_cache/
+
+samples-*
+*.hdf5
+*.h5
+.idea/*
