+++ cnn/unet.py
@@ -1,0 +1,9 @@
+from keras.models import Sequential
+from keras.layers import Dense, Dropout, Flatten
+from keras.layers import Convolution2D
+
+from .base_cnn import BaseCNN
+
+
+class UnetCNN(BaseCNN):
+    pass
